functor_engine_web/backend/requirements.txt
fastapi
uvicorn
networkx
langchain
langchain-google-genai
pydantic
python-dotenv
requests
